rename to makeEditToolIpc
