+++ editor/frontend/src/EditToolIpc.ts
@@ -10,7 +10,7 @@ import { PickAsyncMethods } from "@itwin/core-bentley";
 import { IpcApp } from "@itwin/core-frontend";
 import { BasicManipulationCommandIpc, EditCommandIpc, editorChannel, SolidModelingCommandIpc } from "@itwin/editor-common";
 
-export function makeEditorIpcProxy<K extends EditCommandIpc>(): PickAsyncMethods<K> {
+export function makeEditToolIpc<K extends EditCommandIpc>(): PickAsyncMethods<K> {
   return new Proxy({} as PickAsyncMethods<K>, {
     get(_target, methodName: string) {
       return async (...args: any[]) =>
@@ -19,5 +19,7 @@ export function makeEditorIpcProxy<K extends EditCommandIpc>(): PickAsyncMethods
   });
 }
 
-export const solidModelingIpc = makeEditorIpcProxy<SolidModelingCommandIpc>();
-export const basicManipulationIpc = makeEditorIpcProxy<BasicManipulationCommandIpc>();
+/** Proxy for calling methods in `BasicManipulationCommandIpc` */
+export const basicManipulationIpc = makeEditToolIpc<BasicManipulationCommandIpc>();
+/** Proxy for calling methods in `SolidModelingCommandIpc` */
+export const solidModelingIpc = makeEditToolIpc<SolidModelingCommandIpc>();
