+++ editor/frontend/src/ElementGeometryTool.ts
@@ -21,7 +21,7 @@ import {
 } from "@itwin/editor-common";
 import { computeChordToleranceFromPoint } from "./CreateElementTool";
 import { EditTools } from "./EditTool";
-import { solidModelingIpc } from "./EditorIpcProxy";
+import { solidModelingIpc } from "./EditToolIpc";
 
 /** @alpha */
 export class ElementGeometryGraphicsProvider {
